Move dependencies to shadow-cljs.edn
This is a frontend project which makes keeping project information in deps.edn
and shadow-cljs.edn unnecessarily complex. I also initially thought that
deps.edn would help with separating dev deps but that is apparently not needed
in CLJS [1,2]. This change also makes using Calva straightforward via Jack-in
using the shadow-cljs project type.

[1] https://github.com/thheller/shadow-cljs/issues/166
[2] https://shadow-cljs.github.io/docs/UsersGuide.html#source-paths