Add react-toastify lib and refactor to use PostCSS
Shadow-cljs does not handle importing CSS in a way one can do in Javascript and
defers to other build tools [1]. Therefore, the existing CSS was refactored to
use PostCSS as react-toastify exposes some CSS. Webpack was considered but
making it work with shadow-cljs was messy. Having to use the :external option
for :js-provider and having a double bundle of JS just to get CSS didn't seem
right. PostCSS is more focused on CSS so it seemed to be the most appropriate
tool. Now with PostCSS, shadow-cljs remains the same, and any CSS needed from
node_modules in the future can be added easily without needing to duplicate its
contents elsewhere.

[1]: https://code.thheller.com/blog/shadow-cljs/2020/05/08/how-about-webpack-now.html